Rename ActorInfoExt; move actor formatter to ActorInfoExtensions
